Update status.sh
The >&2 ensures that the message is sent to stderr instead of stdout, so it appears as an error message to the user. When the script checks if it is running with sudo and finds that it's not it displays the error message "Please run this script with sudo." also added for the container line of code.